+++ packages/libs/sdk-component-drivers/src/index.ts
@@ -17,4 +17,5 @@ export * from './UserAttributeDriver';
 export * from './UserAuthMethodDriver';
 export * from './DeviceListDriver';
 export * from './FlowButtonDriver';
+export * from './LastAuthBadgeDriver';
 export * from './TenantSelectorDriver';