+++ packages/sdks/web-component/test/descope-wc.test.ts
@@ -35,6 +35,7 @@ const defaultOptionsValues = {
   oidcLoginHint: null,
   oidcPrompt: null,
   samlIdpStateId: null,
+  wsfedIdpStateId: null,
   samlIdpUsername: null,
   oidcErrorRedirectUri: null,
   oidcResource: null,