+++ packages/widgets/outbound-applications-widget/CHANGELOG.md
@@ -2,6 +2,56 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.37](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.36...outbound-applications-widget-0.2.37) (2026-05-14)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.48.4`
+* `sdk-mixins` updated to version `0.16.2`
+* `sdk-component-drivers` updated to version `0.10.0`
+* `web-component` updated to version `3.62.0`
+## [0.2.36](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.35...outbound-applications-widget-0.2.36) (2026-05-11)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.48.3`
+* `web-component` updated to version `3.61.0`
+## [0.2.35](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.34...outbound-applications-widget-0.2.35) (2026-05-06)
+
+
+### Bug Fixes
+
+* **outboundAppsWidget:** pass tenantId instead of tenant ([#1379](https://github.com/descope/descope-js/issues/1379)) ([c220182](https://github.com/descope/descope-js/commit/c2201822e557e6af49e3f4f2fc21cf4d25a4fee9))
+
+## [0.2.34](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.33...outbound-applications-widget-0.2.34) (2026-04-29)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.48.2`
+* `web-component` updated to version `3.60.0`
+## [0.2.33](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.32...outbound-applications-widget-0.2.33) (2026-04-21)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.48.1`
+* `web-component` updated to version `3.59.2`
+## [0.2.32](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.31...outbound-applications-widget-0.2.32) (2026-04-20)
+
+### Dependency Updates
+
+* `web-component` updated to version `3.59.1`
+## [0.2.31](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.30...outbound-applications-widget-0.2.31) (2026-04-20)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.48.0`
+* `web-component` updated to version `3.59.0`
+## [0.2.30](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.29...outbound-applications-widget-0.2.30) (2026-04-20)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.47.2`
+* `web-component` updated to version `3.58.4`
 ## [0.2.29](https://github.com/descope/descope-js/compare/outbound-applications-widget-0.2.28...outbound-applications-widget-0.2.29) (2026-04-13)
 
 ### Dependency Updates